requirements.txt
async-timeout==4.0.2
click==8.1.3
dash==2.7.1
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
Flask==2.2.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
plotly==5.11.0
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.2
six==1.16.0
tenacity==8.1.0
Werkzeug==2.2.2
